Use workflowUid to create task with workflow
Along with updates to translations-tab, use an optional workflowUid (if
present) passed in by the user to set the workflow on task creation,
rather than hard-coding the specific uid string for the machine
translation workflow.

Dev can pass an optional array of `workflowOptions` in the options
config that's sent to `TranslationsTab`, enabling a select option in
the tab for choosing a workflow and passing in `workflowUid` with the
`createTask` params here.

Example:
```
const myCustomConfig = {
  ...defaultFieldLevelConfig,
  workflowOptions: [
    { workflowUid: '35cdc9b707d9', workflowName: 'Machine Translation (testing)' }
  ]
}
```