Return link to task in Smartling with task
Combined with a parallel update in sanity-translations-tab, this update
includes a link that goes straight to the given task in Smartling, using
the project id and task id. If there is a task, then a link will show
along with the current job to "View Job" so users can easily get to the
associated Smartling job for additional updates or info.